pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sassycode"
version = "0.1.0"
description = "SAST scanner and management interface powered by OpenAI"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.0",
  "openai>=1.40.0",
  "jinja2>=3.1.0",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["pytest", "httpx", "ruff", "black"]

[project.scripts]
sassycode-scanner = "scanner.cli:main"
sassycode-manager = "manager.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["scanner", "manager", "manager.*"]

